requirements.txt
alembic==1.4.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.3
connexion==2.6.0
coverage==5.0.3
docutils==0.15.2
Flask==1.1.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
loguru==0.4.1
Mako==1.1.1
MarkupSafe==1.1.1
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
openapi-spec-validator==0.2.8
packaging==20.1
pluggy==0.13.1
py==1.8.1
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.13
swagger-ui-bundle==0.0.6
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
win32-setctime==1.0.1
WTForms==2.2.1
zipp==3.0.0
